pyproject.toml
[tool.poetry]
name = "brite-ombdb"
version = "0.1.0"
description = ""
authors = ["Ehsan Nouri <ehsanouri@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
alembic = "^1.13.1"
sqlalchemy = "^2.0.27"
aiohttp = "^3.9.3"
uvicorn = "^0.27.1"
fastapi-pagination = "^0.12.15"
pydantic-settings = "^2.1.0"
google-cloud-pubsub = "^2.19.4"
pydantic = {extras = ["dotenv"], version = "*"}
loguru = "^0.7.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
aioresponses = "^0.7.6"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"
aiosqlite = "^0.20.0"
ruff = "^0.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
